+++ YamYamEngine_SOURCE/yaShader.cpp
@@ -13,8 +13,8 @@ namespace ya::graphics
 
 	HRESULT Shader::Load(const std::wstring& path)
 	{
-		int fineNameBeginOffset = path.rfind(L"\\") + 1;
-		int fineNameEndOffset = path.length() - fineNameBeginOffset;
+		size_t fineNameBeginOffset = path.rfind(L"\\") + 1;
+		size_t fineNameEndOffset = path.length() - fineNameBeginOffset;
 		const std::wstring fileName(path.substr(fineNameBeginOffset, fineNameEndOffset));
 
 		if (!Create(eShaderStage::VS, fileName))
